# apps/api/contracts/baselines/legacy-openapi.normalized.json
{
  "components": {
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/social/apple/notifications": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "payload": {
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "payload"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "const": "accepted",
                      "default": "accepted",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/social/attempts": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "flow": {
                    "default": "login",
                    "enum": [
                      "link",
                      "login"
                    ],
                    "type": "string"
                  },
                  "platform": {
                    "enum": [
                      "android",
                      "ios",
                      "web"
                    ],
                    "type": "string"
                  },
                  "provider": {
                    "enum": [
                      "apple",
                      "google",
                      "kakao",
                      "naver"
                    ],
                    "type": "string"
                  },
                  "return_path": {
                    "maxLength": 2048,
                    "type": [
                      "null",
                      "string"
                    ]
                  }
                },
                "required": [
                  "platform",
                  "provider"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "attempt_id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "attempt_secret": {
                      "type": "string"
                    },
                    "expires_at": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "flow": {
                      "enum": [
                        "link",
                        "login"
                      ],
                      "type": "string"
                    },
                    "platform": {
                      "enum": [
                        "android",
                        "ios",
                        "web"
                      ],
                      "type": "string"
                    },
                    "provider": {
                      "enum": [
                        "apple",
                        "google",
                        "kakao",
                        "naver"
                      ],
                      "type": "string"
                    },
                    "return_path": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/social/complete": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "attempt_id": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "attempt_secret": {
                    "maxLength": 256,
                    "minLength": 32,
                    "type": "string"
                  }
                },
                "required": [
                  "attempt_id",
                  "attempt_secret"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "identity": {
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "email_verified": {
                          "type": "boolean"
                        },
                        "linked_at": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "provider": {
                          "enum": [
                            "apple",
                            "google",
                            "kakao",
                            "naver"
                          ],
                          "type": "string"
                        },
                        "status": {
                          "enum": [
                            "active",
                            "revocation_pending",
                            "revoked"
                          ],
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "next_step": {
                      "enum": [
                        "email_required",
                        "legal_consent_required",
                        "profile_completion_required",
                        "role_onboarding_required",
                        "signed_in"
                      ],
                      "type": "string"
                    },
                    "return_path": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/social/identities": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "identities": {
                      "items": {
                        "properties": {
                          "email": {
                            "type": "string"
                          },
                          "email_verified": {
                            "type": "boolean"
                          },
                          "linked_at": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "provider": {
                            "enum": [
                              "apple",
                              "google",
                              "kakao",
                              "naver"
                            ],
                            "type": "string"
                          },
                          "status": {
                            "enum": [
                              "active",
                              "revocation_pending",
                              "revoked"
                            ],
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/social/identities/link-attempts": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "platform": {
                    "enum": [
                      "android",
                      "ios",
                      "web"
                    ],
                    "type": "string"
                  },
                  "provider": {
                    "enum": [
                      "apple",
                      "google",
                      "kakao",
                      "naver"
                    ],
                    "type": "string"
                  },
                  "return_path": {
                    "maxLength": 2048,
                    "type": [
                      "null",
                      "string"
                    ]
                  }
                },
                "required": [
                  "platform",
                  "provider"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "attempt_id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "attempt_secret": {
                      "type": "string"
                    },
                    "expires_at": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "flow": {
                      "enum": [
                        "link",
                        "login"
                      ],
                      "type": "string"
                    },
                    "platform": {
                      "enum": [
                        "android",
                        "ios",
                        "web"
                      ],
                      "type": "string"
                    },
                    "provider": {
                      "enum": [
                        "apple",
                        "google",
                        "kakao",
                        "naver"
                      ],
                      "type": "string"
                    },
                    "return_path": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/social/identities/reconcile": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "identities": {
                      "items": {
                        "properties": {
                          "email": {
                            "type": "string"
                          },
                          "email_verified": {
                            "type": "boolean"
                          },
                          "linked_at": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "provider": {
                            "enum": [
                              "apple",
                              "google",
                              "kakao",
                              "naver"
                            ],
                            "type": "string"
                          },
                          "status": {
                            "enum": [
                              "active",
                              "revocation_pending",
                              "revoked"
                            ],
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "reconciled_at": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "revoked_providers": {
                      "items": {
                        "enum": [
                          "apple",
                          "google",
                          "kakao",
                          "naver"
                        ],
                        "type": "string"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  }
}